+++ packages/lib/src/main.ts
@@ -1,3 +1,5 @@
 export * from '@collectium/core'
 export type * from '@collectium/core'
+export * from '@collectium/check'
+export type * from '@collectium/check'