+++ packages/core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @collectium/core
 
+## 0.2.0
+
+### Minor Changes
+
+- Add cli and check package
+
 ## 0.1.3
 
 ### Patch Changes