+++ packages/lib/CHANGELOG.md
@@ -1,5 +1,18 @@
 # collectium
 
+## 0.2.0
+
+### Minor Changes
+
+- Add cli and check package
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @collectium/core@0.2.0
+  - @collectium/check@0.2.0
+  - @collectium/cli@1.0.0
+
 ## 0.1.3
 
 ### Patch Changes