+++ packages/preset/github/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @collectium/preset-github
 
+## 0.2.0
+
+### Minor Changes
+
+- Add cli and check package
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @collectium/core@0.2.0
+
 ## 0.1.3
 
 ### Patch Changes